+++ parser.h
@@ -38,7 +38,7 @@ struct parsedfile
 /* Functions provided by parser module */
 int read_config(char *, struct parsedfile *);
 int is_local(struct parsedfile *, struct in_addr *);
-int pick_prefix(struct parsedfile *, struct prefixent **, struct in_addr *, unsigned int port);
+int pick_prefix(struct parsedfile *, struct prefixent **, struct in_addr *, uint16_t port);
 int check_prefix(struct parsedfile *config, struct in6_addr * addr);
 char *strsplit(char *separator, char **text, const char *search);
